Instrument database

Instrument database

InstrumentDB takes advantage of the structures Detector and Horn to retrieve information about feed horns and detectors from a YAML file. There are a set of YAML files containing the default configuration for the STRIP instrument in the repository.

Quick introduction

The following example initializes an object of type InstrumentDB with the values referred to the standard STRIP instrument:

julia> db = InstrumentDB()
InstrumentDB(49 horns, 66 detectors)

As db is a struct, its field can be accessed with the usual dot notation. The two fields in db are focalplane and detectors. They are both dictionaries, associating horn names to Horn objects and detectors IDs to Detector objects, respectively:

julia> db.focalplane["I0"]
Horn I0, module 0 (indigo):
    Orientation: [0.0000, -0.0000, 1.0000]
    FWHM (X/Y): 0.3470 × 0.3440 °
    Spillover: 0.016704 (main), 0.013964 (sub)
    Cross-polarization: 44.53 dB
    Directivity: 54.76 dBi
    Ellipticity: 1.0080

julia> db.detectors[2]
Detector STRIP02 (Q band):
    Center frequency: 43.39 ± 0.13 GHz
    Bandwidth: 7.53 ± 0.39 GHz
    Noise temperature: 31.5 ± 3.92 K
    Knee frequency: 90.0 ± 40.0 mHz (Q), 70.0 ± 40.0 mHz (U)

The structure Detector is complex, as it is built over three other structures:

BandshapeInfo

SpectrumInfo

NoiseTemperatureInfo

All these structures know how to show themselves on the REPL:

julia> db.detectors[2].bandshape
Bandshape:
    Center frequency: 43.39 ± 0.13 GHz
    Bandwidth: 7.53 ± 0.39 GHz
    Frequency range: [38.00, 50.00] GHz (121 points)
    Test ID: 355
    Analysis ID: 89

julia> db.detectors[2].spectrum
Noise spectrum:
    Slope: I = 0.8585 ± 0.0000, Q = 0.6420 ± 0.0030, U = 0.3680 ± 0.0030
    Knee frequency: I = 0.0 ± 0.0 mHz, Q = 90.0 ± 40.0 mHz, U = 70.0 ± 40.0 mHz
    White noise: Q = 5.0 ± 1.0 mK^2 Hz, U = 2.6 ± 0.6 mK^2 Hz
    Test ID: 357
    Analysis ID: 15

julia> db.detectors[2].tnoise
Noise temperature:
    Tnoise: 31.5 ± 3.9 K
    Estimates: [26.7, 27.1, 27.6, 28.2, 30.1, 31.1, 32.0, 33.6, 34.0, 35.1, 36.2, 39.7] K
    Test IDs: []
    Analysis IDs: [105, 106]

For more information about the fields in the structures listed above, as well as their meaning, keep reading.

Structures

Stripeline.InstrumentDB — Type.

STRIP instrument database

The "database" contains information about feed horns and polarimeters:

The field focalplane is a dictionary (mapping) associating the string identifying a horn (e.g., I0) with a Horn structure;

The field detectors is a dictionary associating the ID of the polarimeter (e.g., 2 stands for STRIP02) with a Detector structure.

You should usually create an object of this kind using the default constructor, which parses a set of YAML files containing the real parameters of the instrument.

Examples

julia> db = InstrumentDB();

julia> print("Number of horns in the database: $(length(keys(db.focalplane)))")
Number of horns in the database: 49

julia> print("Number of polarimeters in the database: $(length(keys(db.detectors)))")
Number of polarimeters in the database: 66

source

Stripeline.Horn — Type.

Information about a STRIP horn

This structure holds a number of parameters relative to each feed horn in the STRIP focal plane.

You should initialize Horn objects via the InstrumentDB constructor, which loads their definition from a STRIP instrument database in YAML format.

Field	Type	Meaning

name	String	Name of the horn, e.g., I0

id	Int	Unique number of the horn, starting from 1

polid	Int	Unique ID of the polarimeter associated with the horn

moduleid	Int	Number of the horn within the module, from 0 to 6

color	String	Name of the color associated with the module

orientation	Array{Float64}	3D vector containing the orientation of the horn in the sky

fwhm_x_deg	Float64	FWHM of the beam along the X axis, in degrees

fwhm_y_deg	Float64	FWHM of the beam along the Y axis, in degrees

main_spillover	Float64	Main reflector spillover

sub_spillover	Float64	Sub-reflector spillover

xpd_db	Float64	Cross-polarization, in dB

directivity_dbi	Float64	Directivity, in dBi

ellipticity	Float64	Ellipticity

source

Stripeline.Detector — Type.

Information about a STRIP detector

This structure holds information about a STRIP polarimeter.

You should initialize Detector objects via the InstrumentDB constructor, which loads their definition from a local STRIP instrument database.

Field	Type	Meaning

id	Int	Integer ID of the polarimeter, e.g., 2 for STRIP02

name	String	Full name of the polarimeter, e.g., STRIP02

band	String	Band: it can either be Q or W

bandshape	BandshapeInfo	Information about the bandpass response

spectrum	SpectrumInfo	Information about the noise spectrum (white noise and 1/f noise)

tnoise	NoiseTemperatureInfo	Information about the noise temperature

source

Stripeline.BandshapeInfo — Type.

BandshapeInfo

Information about the spectral band response of a polarimeter.

Field	Type	Meaning

center_frequency_hz	Float64	Estimate for the center frequency, in Hz

center_frequency_err_hz	Float64	Estimated error on the center frequency, in Hz

bandwidth_hz	Float64	Estimated bandwidth, in Hz

bandwidth_err_hz	Float64	Estimated error on the bandwidth, in Hz

lowest_frequency_hz	Float64	Lowest frequency of the bandshape in response, in Hz

highest_frequency_hz	Float64	Highest frequency of the bandshape in response, in Hz

num_of_frequencies	Int	Number of samples in response

response	Array{1,Float64}	Profile of the bandshape (pure numbers)

test_id	Int	ID of the unit-level test used to characterize the bandshape

analysis_id	Int	ID of the unit-level analysis used to characterize the bandshape

source

Stripeline.SpectrumInfo — Type.

SpectrumInfo

Information about the noise spectrum of the output of a polarimeter.

Field	Type	Meaning

slope_i	Float64	The slope ($\alpha$) of the 1/f component of the noise in the I signal

slope_i_err	Float64	Error associated with the value of slope_i

slope_q	Float64	Same as slope_i, but for the Q signal

slope_q_err	Float64	Error associated with the value of slope_q

slope_u	Float64	Same as slope_i, but for the U signal

slope_u_err	Float64	Error associated with the value of slope_u

fknee_i_hz	Float64	Knee frequency of the I signal, in Hz

fknee_i_err_hz	Float64	Error associated with the value of fknee_i_hz

fknee_q_hz	Float64	Knee frequency of the Q signal, in Hz

fknee_q_err_hz	Float64	Error associated with the value of fknee_q_hz

fknee_u_hz	Float64	Knee frequency of the U signal, in Hz

fknee_u_err_hz	Float64	Error associated with the value of fknee_u_hz

wn_i_k2_hz	Float64	White noise level for the I signal, in K^2 Hz

wn_i_err_k2_hz	Float64	Error associated with the value of wn_i_k2_hz

wn_q_k2_hz	Float64	White noise level for the Q signal, in K^2 Hz

wn_q_err_k2_hz	Float64	Error associated with the value of wn_q_k2_hz

wn_u_k2_hz	Float64	White noise level for the U signal, in K^2 Hz

wn_u_err_k2_hz	Float64	Error associated with the value of wn_u_k2_hz

test_id	Int	ID of the unit-level test used to characterize the bandshape

analysis_id	Int	ID of the unit-level analysis used to characterize the bandshape

source

Stripeline.NoiseTemperatureInfo — Type.

NoiseTemperatureInfo

Information about the noise temperature of a polarimeter. This structure is used for the field tnoise of the Detector struct.

Field	Type	Meaning

tnoise_k	Float64	Noise temperature computed from tnoise_values_k, in K

tnoise_err_k	Float64	Error associated with tnoise_k, computed from tnoise_values_k

test_ids	Array{Int,1}	List of unit-level test IDs used to estimate the noise temperature

analysis_ids	Array{Int,1}	List of unit-level analysis report IDs used to estimate the noise temperature

values_k	Array{Float64,1}	List of noise temperatures estimated from the tests

source

Loading custom databases

It is not needed to load the default instrument database, as Stripeline provides a number of additional functions to build mock databases from dictionaries.

Stripeline.defaultdbfolder — Function.

defaultdbfolder()

Return a string containing the (local) full path to the YAML files containing the reference instrument DB.

source

Stripeline.parsefpdict — Function.

parsefpdict(fpdict)

Return a dictionary associating an horn name (e.g., I0) to a Horn object containing information about some horn in the STRIP focal plane. The information are parsed from fpdict, which should be a dictionary loaded from a YAML file. The default YAML file to be used is located in the folder returned by defaultdbfolder and is usually named strip_focal_plane.yaml

source

Stripeline.parsedetdict — Function.

parsedetdict(detdict)

Return a dictionary associating an integer number to a Detector object containing information about the STRIP detector with the corresponding number. The information are parsed from detdict, which should be a dictionary loaded from a YAML file. The default YAML file to be used is located in the folder returned by defaultdbfolder and is usually named strip_detectors.yaml

source