Random numbers

Random number generators

Although Julia already implements a number of pseudo-random number generators, Stripeline implements its own generators. Strip re-implements the same generators used in the pipeline of the Planck/LFI instrument, which provided several types of distributions:

Uniform distribution (Flat128RNG), with period 2^128;

Gaussian distribution (GaussRNG);

\[1/f^2\]

distribution (Oof2RNG);

\[1/f^α\]

distribution, with $α < 2$ (OofRNG).

Each generator but Flat128RNG uses a simpler generator internally. This generator must sample from a given distribution, but it does not need to be a generator provided by Stripeline. For instance, the Gaussian generator GaussRNG employs an uniform generator, which can either be Flat128RNG or one of the generators provided by Julia like MersenneTwister. For instance, here is an example which shows how to use Flat128RNG:

julia> gauss1 = GaussRNG(initflatrng128(1234))
Stripeline.GaussRNG(Stripeline.Flat128RNG(UInt32[0x5bf4963e, 0x7f2bb362, 0x8dfeefd3, 0xef95f09a]), true, 0)

julia> print([randn(gauss1) for i in 1:4])
[-0.635931, -0.881716, -0.0869356, -0.403257]

We use initflatrng128, as it creates a Flat128RNG object with some sensible defaults (specifically, it is configured to produce the same sequence of random numbers as the ones produced by the Planck/HFI pipeline, if the seeds are the same). And here is the same example, using a MersenneTwister generator:

julia> gauss2 = GaussRNG(MersenneTwister(1234))
Stripeline.GaussRNG(MersenneTwister(UInt32[0x000004d2], Base.dSFMT.DSFMT_state(Int32[-1393240018, 1073611148, 45497681, 1072875908, 436273599, 1073674613, -2043716458, 1073445557, -254908435, 1072827086  …  -599655111, 1073144102, 367655457, 1072985259, -1278750689, 1018350124, -597141475, 249849711, 382, 0]), [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0  …  0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], 382), true, 0)

julia> print([randn(gauss2) for i in 1:4])
[1.43347, 0.488099, -1.41027, 2.34031]

Of course, the numbers are different. They are however drawn from the same distribution (Gaussian curve with mean 0 and σ=1).

Uniform generator

Stripeline.Flat128RNG — Type.

Flat128RNG

State of the base-128 uniform random generator. Initialize this using the function initflatrng128.

source

Stripeline.initflatrng128 — Function.

initflatrng128(xstart = 123456789, ystart = 362436069, zstart = 521288629, wstart = 88675123)

Initialize a flat random number generator with period 2^128. To draw random numbers, use the Base.rand function as usual.

Example:

rng = initflatrng128()
print([rand(rng) for i in 1:4])

source

Gaussian generator

Stripeline.GaussRNG — Type.

GaussRNG(flatrng::AbstractRNG)
GaussRNG(seed=0)

Initialize a Gaussian RNG. The parameter flatrng must be a uniform RNG. If a seed is used, then a MersenneTwister RNG is used.

source

$1/f^2$ generator

Stripeline.Oof2RNG — Type.

Oof2RNG(normrng, fmin, fknee, fsample)

Create a Oof2RNG RNG object. It requires a gaussian RNG generator in normrng (use GaussRNG), the minimum frequency (longest period) in fmin, the knee frequency and the sampling frequency. The measure unit of the three frequencies must be the same (e.g., Hz).

Use randoof2 to draw samples from a Oof2RNG object, as in the following example:

rng = Oof2RNG(GaussRNG(), 1e-3, 1.0, 1e2)
print([randoof2(rng) for i in 1:4])

source

$1/f^α$ generator (with $α < 2$)

Stripeline.OofRNG — Type.

OofRNG(normrng, fmin, fknee, fsample)

Create a OofRNG RNG object. It requires a gaussian RNG generator in normrng (use GaussRNG), the slope α of the noise in slope, the minimum frequency (longest period) in fmin, the knee frequency and the sampling frequency. The measure unit of the three frequencies must be the same (e.g., Hz).

Use randoof to draw samples from a OofRNG object, as in the following example:

rng = OofRNG(GaussRNG(), 1.5, 1e-3, 1.0, 1e2)
print([randoof(rng) for i in 1:4])

source