Basic functions

Measure unit conversion

It is often useful to convert measurements between thermodynamic temperatures and Rayleigh-Jeans temperatures. The following functions implement this kind of conversion. Note that there are two families of functions:

Functions that convert absolute measurements: t_to_trj, trj_to_t;

Functions that convert sensitivities (i.e., small fluctuations around an absolute value): deltat_to_deltatrj, deltatrj_to_deltat

The function sensitivity_tant computes the overall sensitivity of a set of polarimeters, using information from The Strip Instrument Database.

Stripeline.sensitivity_tant — Function

sensitivity_tant(db::InstrumentDB, load_tant; modules = Set([0, 1, 2, 3, 4, 5, 6]))

Calculate the white-noise sensitivity of an array of detectors, measured in K⋅√s, given some antenna temperature for the load. The result takes in account only those horns belonging to the modules listed in the keyword modules (the W-band horns belong to module -1). By default, only the Q-band modules are considered.

The result assumes the radiometer equation: $σ⋅√τ = \frac{T_{sys}}{√2β}$, where T_{sys} is the system temperature, β is the bandwidth, and τ is the acquisition time. The factor 2 comes from the way Strip polarimeters operate. The system temperature is assumed to be the noise temperature of each detector, plus the term load_tant, which should take into account all the other sources of power entering the system (e.g., telescope, atmosphere, etc.). The term load_tant should be expressed as an antenna temperature.

Stripeline.t_to_trj — Function

t_to_trj(temperature_k, nu_hz)

Convert a thermodynamic temperature (in K) into a Rayleigh-Jeans temperature, given some specified frequency nu_hz (in Hz).

See also trj_to_t for the inverse transformation.

Stripeline.trj_to_t — Function

trj_to_t(temperature_k, nu_hz)

Convert a Rayleigh-Jeans temperature (in K) into a thermodynamic temperature, given some specified frequency nu_hz (in Hz).

See also t_to_trj for the inverse transformation.

Stripeline.deltat_to_deltatrj — Function

deltat_to_deltatrj(temperature_k, deltat_k, nu_hz)

Convert a small temperature fluctuation deltat_k around temperature temperature_k from thermodynamic temperature to Rayleigh-Jeans (RJ) temperature. This function can be used to convert sensitivities expressed as thermodynamic temperatures in RJ sensitivities.

See also deltatrj_to_deltat for the inverse function.

Stripeline.deltatrj_to_deltat — Function

deltat_to_deltatrj(temperature_k, deltat_k, nu_hz)

Convert a small temperature fluctuation deltat_k around temperature temperature_k from Rayleigh-Jeans (RJ) temperature to thermodynamic temperature. This function can be used to convert sensitivities expressed as RJ temperatures in thermodynamic sensitivities.

See also deltatrj_to_deltat for the inverse function.