Map making

Map-making functions

Stripeline implements a few functions to produce maps out of time-ordered data.

Stripeline.tod2map_mpi — Function.

tod2map(pix_idx, tod, num_of_pixels, comm) -> binned_map
tod2map(pix_idx, tod, num_of_pixels, data_properties, comm) -> binned_map

This function creates a binned map from the time-ordered data kept in the array tod, assuming that each sample is observing the pixel whose index is in pix_idx. The parameter num_of_pixels contains the number of pixels in the Healpix map, and it is used as an upper bound for the values in pix_idx. The parameter comm must be a MPI communicator, or missing if you are not using MPI. This is a MPI based function: each MPI process computes a map from its available data. All partial maps are then combined together with MPI.allreduce. The function returns an array containing the binned map.

If Array of structures data_properties_structis passed to the function, the output binned map will be a weighted binned map. Each sample will be weighted according to the inverse white noise variance σ^2 of the corrisponding polarimeter. In this way, the less noisy polarimeters will count more in the estimation of the map.

Requirements

The length of the arrays pix_idx and tod must be the same

source

Stripeline.baseline2map_mpi — Function.

baseline2map_mpi(pix_idx, baselines, baseline_lengths, num_of_pixels, comm)-> noise_map
baseline2map_mpi(pix_idx, baselines, num_of_pixels, data_properties, comm) -> noise_map

This is a MPI based function: each MPI process computes a map from its available data. All partial maps are then combined together with MPI.allreduce. The function returns an array containing the binned map.

If Array of structures data_properties_structis passed to the function (instead of baseline_lengths) the output binned map will be a weighted binned map. Each sample will be weighted according to the inverse white noise variance σ^2 of the corrisponding polarimeter. In this way, the less noisy polarimeters will count more in the estimation of the map.

Requirements

The length of baselines and baseline_lengths must be the same;

The value sum(baseline_lengths) must be the same as the length of pix_idx.

source

Stripeline.destripe — Function.

destripe(pix_idx, tod, num_of_pixels, data_properties, rank, comm; threshold = 1e-9, max_iter = 10000) -> (pixels, baselines)

This MPI based function creates a map from a TOD and removes both 1/f and white noise, using the destriping technique.

The parameters passed to the function have the following meaning:

pix_idx: array containing the indices of the pixels visited by the instrument

tod: the values measured by the polarimeters for each pixel (either I, Q, or U)

num_of_pixels: the number of pixels in the map to be produced. This is used as an upper limit for the values in pix_idx

data_properties: an array of structures data_properties_struct, holding information on each simulated polarimeter noise level, number and length of 1/f baselines and total number of samples. It can be obtained by using function get_data_properties.

rank: the rank of the current MPI process

comm: the MPI communicator object.

The following arguments are optional:

threshold is used by the conjugated-gradient algorithm. When the residual error of the iteration goes below this value, the iteration stops. The smaller the value, the more accurate the solution.

max_iter is the maximum number of iterations to be executed in the conjugated-gradient algorithm. If the algorithm does not converge within this number of iterations, the process will quit without having reached the convergence threshold (see the threshold keyword above).

The function returns a 2-tuple containing the destriped map itself (Array{T,1}, where T is the base type used for the TOD in the argumenttod) and an array containing the baselines. Since it is not granted that the sequence of convergence parameters of the conjugate gradient is monotonically decreasing, the code keeps the lowest value of them and the corresponding array of baselines. If the loop ends because the maximum number of iterations has been reached, this is the configuration that will be returned to the caller.

Remarks

The length of the arrays pix_idx and tod must be the same;

If you are not using MPI, pass missing to the comm parameter.

source