Scanning strategy

Simulating the scanning strategy

Stripeline.genskypointings — Function.

genskypointings(t_start, t_stop, dirs; latitude_deg=0.0, longitude_deg=0.0, height_m=0.0)

Project a set of pointings in the sky.

The parameter t_start and t_start must be two strings which tells the exact UTC date and time of the observation in "iso" format. The parameter dirs must be a N×2 array containing the observed directions expressed in colatitude and the longitude. The keywords latitude_deg, longitude_deg and height_m should contain the latitude (in degrees, N is positive), the longitude (in degrees, counterclockwise is positive) and the height (in meters) of the location where the observation is made.

Return a 2-tuple containing the observed directions projected in sky coordinates (a N×2 array containing the Right ascension and the Declination, in radians) at each time step. Directions are expressed in ICRS coordinates.

Example:

genskypointings("2019-01-01 00:00:00", "2020-01-25 14:52:10.05", dirs)

source

Stripeline.genpointings — Function.

genpointings(wheelanglesfn, dir, timerange_s; latitude_deg=0.0)

Generate a set of pointings for some STRIP detector. The parameter wheelanglesfn must be a function which takes as input a time in seconds and returns a 3-tuple containing the angles (in radians) of the three motors:

The boresight motor

The altitude motor

The ground motor

The parameter dir must be a normalized vector which tells the pointing direction of the beam (boresight is [0, 0, 1]). The parameter timerange_s is either a range or a vector which specifies at which times (in second) the pointings should be computed. The keyword latitude_deg should contain the latitude (in degrees, N is positive) of the location where the observation is made.

Return a 2-tuple containing the directions (a N×2 array containing the colatitude and the longitude) and the polarization angles at each time step. Directions are expressed in equatorial coordinates.

Example:

genpointings([0, 0, 1], 0:0.1:1) do time_s
    # Boresight motor keeps a constant angle equal to 0°
    # Altitude motor remains at 20° from the Zenith
    # Ground motor spins at 1 RPM
    return (0.0, deg2rad(20.0), timetorotang(time_s, 1))
end

source

Stripeline.timetorotang — Function.

timetorotang(time, rpm)

Convert a time into a rotation angle, given the number of rotations per minute. The time should be expressed in seconds. The return value is in radians. time can either be a scalar or a vector.

source