Scanning strategy

Simulating the scanning strategy

Stripeline.telescopetoground — Function.

telescopetoground(wheelanglesfn, time_s)

Return a quaternion of type Quaternion{Float64} representing the coordinate transform from the focal plane to the ground of the telescope. The parameter wheelanglesfn must be a function which takes as input a time, time_s, in seconds, and it must return a 3-tuple containing the angles of the following motors, measured in radians:

The boresight motor (rotation around the $z$ axis, counterclockwise)

The altitude motor (rotation around the $y$ axis, counterclockwise)

The ground motor (rotation around the $z$ axis, clockwise: N→E→S→W)

Example

telescopetoground(3600.0) do
    # Boresight motor keeps a constant angle equal to 0°
    # Altitude motor remains at 20° from the Zenith
    # Ground motor spins at 1 RPM
    (0.0, deg2rad(20.0), timetorotang(time_s, 1))
end

source

Stripeline.groundtoearth — Function.

groundtoearth(groundq, time_s, latitude_deg; day_duration_s=86400.0)

Return a quaternion of type Quaternion{Float64} representing the coordinate transformation from the ground of the telescope to the Equatorial coordinate system. The parameter groundq must be a quaternion describing the coordinate transformation from the focal plane of the telescope to the ground. The parameter time_s must be a time in seconds, and latitude_deg is the latitude (in degrees, N is positive) of the location where the observation is made.

The keyword day_duration_s specifies the length of a sidereal day in seconds.

source

Stripeline.genpointings — Function.

genpointings!(wheelanglesfn, beam_dir, timerange_s, dirs, psi; 
              polaxis = Float64[1.0, 0.0, 0.0],
              latitude_deg = TENERIFE_LATITUDE_DEG, 
              ground = false)
genpointings(wheelanglesfn, beam_dir, timerange_s; 
             polaxis = Float64[1.0, 0.0, 0.0],
             latitude_deg = TENERIFE_LATITUDE_DEG,
             ground = false)
genpointings!(wheelanglesfn, beam_dir, timerange_s, t_start, dirs, psi;
              polaxis = Float64[1.0, 0.0, 0.0],
              latitude_deg = TENERIFE_LATITUDE_DEG, 
              longitude_deg = TENERIFE_LONGITUDE_DEG,
              height_m = TENERIFE_HEIGHT_M,
              precession = true,
              nutation = true, 
              aberration = true, 
              refraction = true)
genpointings(wheelanglesfn, beam_dir, timerange_s, t_start;
             polaxis=Float64[1.0, 0.0, 0.0],
             latitude_deg=TENERIFE_LATITUDE_DEG, 
             longitude_deg=TENERIFE_LONGITUDE_DEG,
             height_m=TENERIFE_HEIGHT_M,
             precession = true,
             nutation = true, 
             aberration = true, 
             refraction = true)

Generate a set of pointing directions for a STRIP detector. Each function is provided in two flavours: the ones ending with ! save the results in the last two parameters dirs and psi, while the others automatically allocate memory and return their results as a pair (dirs, psi).

The parameter wheelanglesfn must be a function which takes as input a time in seconds and returns a 3-tuple containing the angles (in radians) of the three motors:

The boresight motor (rotation around the $z$ axis, counterclockwise)

The altitude motor (rotation around the $y$ axis, counterclockwise)

The ground motor (rotation around the $z$ axis, clockwise: N→E→S→W)

The meaning of the parameters/keywords is the following:

beam_dir specifies the pointing direction of the mean (the boresight is [0, 0, 1]). It must be normalized.

timerange_s is an enumerable type that specifies at which times (in seconds) pointings must be computed.

t_start is a DateTime which tells the exact UTC date and time of the observation.

latitude_deg is the latitude of the location where the observation is made (in degrees, North is positive).

ground is a Boolean: if true, the function will return a 4-tuple containing the colatitude and longitude measured in Equatorial coordinates (columns 1 and 2) and in ground coordinates (columns 3 and 4). If false, only the Equatorial coordinates are computed.

polaxis is the polarization axis; it must be normalized.

precession: if true (the default), the Earth's precession is taken into account.

nutation: if true (the default), the Earth's nutation is taken into account.

aberration: if true (the default), stellar aberration is taken into account.

refraction: if true, refraction corrections are taken into account. As these corrections are only valid for optical wavelengths, the default is false.

Return values

If t_start is not provided, the function genpointings returns a 2-tuple containing the sky directions (a N×2 array containing declination and right ascension, in Equatorial coordinates) and the polarization angle for each time step. The function genpointings! sets the values in the last two parameters dirs and psi.

If t_start is provided, the function genpointings returns a 2-tuple (4-tuple) containing the directions (a N×2 or Nx4 array containing the colatitude and the longitude) and the polarization angles at each time step; genpointings! works as above.

Examples

Here is an example using the form without t_start:

dir, psi = genpointings([0, 0, 1], 0:0.1:1) do time_s
    # Boresight motor keeps a constant angle equal to 0°
    # Altitude motor remains at 20° from the Zenith
    # Ground motor spins at 1 RPM
    (0.0, deg2rad(20.0), timetorotang(time_s, 1))
end

And here is an example using t_start; unlike the previous example, we use a lambda function instead of the do...end notation.

import Dates

dirs, psi = genpointings(time_s -> (0, deg2rad(20),
                                    timetorotang(time_s, 1)),
                         [0, 0, 1],
                         0:0.1:1,
                         Dates.DateTime(2022, 01, 01, 0, 0, 0),
                         latitude_deg=10.0,
                         longitude_deg=20.0,
                         height_m = 1000) do time_s

source

Stripeline.timetorotang — Function.

timetorotang(time, rpm)

Convert a time into a rotation angle, given the number of rotations per minute. The time should be expressed in seconds. The return value is in radians. time can either be a scalar or a vector.

source

Stripeline.northdir — Function.

northdir(θ, ϕ)
eastdir(θ, ϕ)

Compute the North/East versor for a vector. The North for a vector v is -dv/dθ, as θ is the colatitude and moves along the meridian, and the East is dv/dϕ.

Examples

julia> northdir(π/2, 0) ≈ [0, 0, 1]
true
julia> eastdir(π/2, 0) ≈ [1, 0, 0]
true

source

Stripeline.eastdir — Function.

northdir(θ, ϕ)
eastdir(θ, ϕ)

Compute the North/East versor for a vector. The North for a vector v is -dv/dθ, as θ is the colatitude and moves along the meridian, and the East is dv/dϕ.

Examples

julia> northdir(π/2, 0) ≈ [0, 0, 1]
true
julia> eastdir(π/2, 0) ≈ [1, 0, 0]
true

source

Stripeline.polarizationangle — Function.

polarizationangle(northdir, eastdir, poldir)

Calculate the polarization angle projected in the sky in IAU conventions. The parameters northdir and eastdir must be versors that point towards the North and East, respectively; poldir must be a versor that identify the polarization direction projected in the sky. The return value is in radians, and it is zero if the polarization angles points toward East, π/2 if it points toward North, etc.

Examples

julia> polarizationangle([0, 0, 1], [0, 1, 0], [0, 1, 0])
0.0
julia> polarizationangle([0, 0, 1], [0, 1, 0], [0, 0, 1]) |> rad2deg
90.0
julia> polarizationangle([0, 0, 1], [0, 1, 0], [0, 0, -1]) |> rad2deg
-90.0

source