Scanning strategy

Simulating the scanning strategy

Stripeline.telescopetoground — Function.

telescopetoground(wheelanglesfn, time_s)

Return a quaternion of type Quaternion{Float64} representing the coordinate transform from the focal plane to the ground of the telescope. The parameter wheelanglesfn must be a function which takes as input a time, time_s, in seconds, and it must return a 3-tuple containing the angles of the following motors, measured in radians:

The boresight motor (rotation around the $z$ axis, counterclockwise)

The altitude motor (rotation around the $y$ axis, counterclockwise)

The ground motor (rotation around the $z$ axis, clockwise: N→E→S→W)

Example

telescopetoground(3600.0) do
    # Boresight motor keeps a constant angle equal to 0°
    # Altitude motor remains at 20° from the Zenith
    # Ground motor spins at 1 RPM
    (0.0, deg2rad(20.0), timetorotang(time_s, 1))
end

source

Stripeline.groundtoearth — Function.

groundtoearth(groundq, time_s, latitude_deg; day_duration_s=86400.0)

Return a quaternion of type Quaternion{Float64} representing the coordinate transformation from the ground of the telescope to the Equatorial coordinate system. The parameter groundq must be a quaternion describing the coordinate transformation from the focal plane of the telescope to the ground. The parameter time_s must be a time in seconds, and latitude_deg is the latitude (in degrees, N is positive) of the location where the observation is made.

The keyword day_duration_s specifies the length of a sidereal day in seconds.

source

Stripeline.genpointings — Function.

genpointings(wheelanglesfn, dir, timerange_s; 
             polaxis=Float64[1.0, 0.0, 0.0],
             latitude_deg=0.0, ground=false)
genpointings(wheelanglesfn, dir, timerange_s, t_start;
             polaxis=Float64[1.0, 0.0, 0.0],
             latitude_deg=0.0, longitude_deg=0.0, height_m=0.0)

Generate a set of pointings for a STRIP detector. The parameter wheelanglesfn must be a function which takes as input a time in seconds and returns a 3-tuple containing the angles (in radians) of the three motors:

The boresight motor (rotation around the $z$ axis, counterclockwise)

The altitude motor (rotation around the $y$ axis, counterclockwise)

The ground motor (rotation around the $z$ axis, clockwise: N→E→S→W)

The meaning of the parameters/keywords is the following:

dir specifies the pointing direction of the mean (the boresight is [0, 0, 1]). It must be normalized.

timerange_s is an enumerable type that specifies at which times (in seconds) pointings must be computed.

t_start is a DateTime which tells the exact UTC date and time of the observation.

latitude_deg is the latitude of the location where the observation is made (in degrees, North is positive).

ground is a Boolean: if true, the function will return a 4-tuple containing the colatitude and longitude measured in Equatorial coordinates (columns 1 and 2) and in ground coordinates (columns 3 and 4). If false, only the Equatorial coordinates are computed.

polaxis is the polarization axis; it must be normalized.

precession is a boolen parameter that allow to include the Earth's precession effect; default is true.

nutation is a boolen parameter that allow to include the Earth's nutation effect; default is true.

aberration is a boolen parameter that allow to include the stellar aberration effect; default is true.

refract is a boolen parameter that allow to include the refraction correction ; default is true.

Return values

In the first form, the function returns a 2-tuple containing the sky directions (a N×2 array containing declination and right ascension, in Equatorial coordinates) and the polarization angle for each time step.

In the second form, the function returns a 2-tuple (4-tuple) containing the directions (a N×2 or Nx4 array containing the colatitude and the longitude) and the polarization angles at each time step.

Examples

Here is an example of the first kind:

dir, psi = genpointings([0, 0, 1], 0:0.1:1) do time_s
    # Boresight motor keeps a constant angle equal to 0°
    # Altitude motor remains at 20° from the Zenith
    # Ground motor spins at 1 RPM
    return (0.0, deg2rad(20.0), timetorotang(time_s, 1))
end

And here is an example of the second kind. Note that, unlike the previous example, we use a lambda function instead of the do...end notation.

import Dates

dirs, psi = genpointings(time_s -> (0, deg2rad(20),
                                    timetorotang(time_s, 1)),
                         [0, 0, 1],
                         0:0.1:1,
                         Dates.DateTime(2019, 01, 01, 0, 0, 0),
                         latitude_deg=10.0,
                         longitude_deg=20.0,
                         height_m = 1000) do time_s

source

Stripeline.timetorotang — Function.

timetorotang(time, rpm)

Convert a time into a rotation angle, given the number of rotations per minute. The time should be expressed in seconds. The return value is in radians. time can either be a scalar or a vector.

source